+++ internal/terminal/terminal_check_aix.go
@@ -1,6 +1,6 @@
 // +build aix
 
-package logrus
+package terminal
 
 import (
 	"golang.org/x/sys/unix"
