+++ internal/terminal/terminal_check_appengine.go
@@ -6,6 +6,6 @@ import (
 	"io"
 )
 
-func checkIfTerminal(w io.Writer) bool {
+func IsTerminal(w io.Writer) bool {
 	return true
 }
