+++ internal/terminal/terminal_check_linux.go
@@ -1,6 +1,6 @@
 // +build linux
 
-package logrus
+package terminal
 
 import "golang.org/x/sys/unix"
