+++ internal/terminal/terminal_notwindows.go
@@ -4,5 +4,5 @@ package logrus
 
 import "io"
 
-func initTerminal(w io.Writer) {
+func InitTerminal(w io.Writer) {
 }
